Data_Loader/requirements.txt
greenlet==2.0.1
numpy==1.23.4
pandas==1.5.1
psycopg==3.1.4
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.4.0
